Use double for storing time instead of float Float precision may be not enough after some time of operating

+++ flix/cli.ino
@@ -8,7 +8,7 @@
 
 extern const int MOTOR_REAR_LEFT, MOTOR_REAR_RIGHT, MOTOR_FRONT_RIGHT, MOTOR_FRONT_LEFT;
 extern float loopRate, dt;
-extern float t;
+extern double t;
 extern int rollChannel, pitchChannel, throttleChannel, yawChannel, armedChannel, modeChannel;
 
 const char* motd =
