+++ flix/flix.ino
@@ -10,7 +10,7 @@
 #define SERIAL_BAUDRATE 115200
 #define WIFI_ENABLED 1
 
-float t = NAN; // current step time, s
+double t = NAN; // current step time, s
 float dt; // time delta from previous step, s
 int16_t channels[16]; // raw rc channels
 float controls[16]; // normalized controls in range [-1..1] ([0..1] for throttle)
